requirements.txt
# =============================================================================
# Payment Stream Pipeline — Python Dependencies
# =============================================================================

# Core
pyspark==3.5.1
delta-spark==3.1.0

# Data generation
faker==24.3.0
pydantic==2.6.4
numpy==1.26.4
pandas==2.2.1

# Kafka (optional streaming source)
confluent-kafka==2.3.0
kafka-python==2.0.2

# Airflow
apache-airflow==2.9.1
apache-airflow-providers-apache-spark==4.8.0
apache-airflow-providers-slack==8.5.0

# Data quality
great-expectations==0.18.12

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
pyarrow==15.0.2
boto3==1.34.69          # S3-compatible object storage support
google-cloud-storage==2.16.0

# Observability
structlog==24.1.0
prometheus-client==0.20.0

# Testing
pytest==8.1.1
pytest-cov==5.0.0
chispa==0.9.2           # PySpark DataFrame equality assertions

# Dev
black==24.3.0
ruff==0.3.4
mypy==1.9.0
pre-commit==3.7.0
